refactor(app-router): isolate visible commit approval authority
