+++ packages/vinext/src/server/app-browser-state.ts
@@ -66,7 +66,7 @@ export type AppRouterAction = {
   type: "navigate" | "replace" | "traverse";
 };
 
-type PendingNavigationCommit = {
+export type PendingNavigationCommit = {
   action: AppRouterAction;
   interceptionContext: string | null;
   previousNextUrl: string | null;
@@ -74,48 +74,9 @@ type PendingNavigationCommit = {
   routeId: string;
 };
 
-type PendingNavigationCommitDisposition = "dispatch" | "hard-navigate" | "skip";
+export type PendingNavigationCommitDisposition = "dispatch" | "hard-navigate" | "skip";
 type PendingNavigationCommitDispositionDecision = {
   disposition: PendingNavigationCommitDisposition;
-  trace: NavigationTrace;
-};
-type VisibleCommitDecision = {
-  disposition: "commit";
-  trace: NavigationTrace;
-};
-type HardNavigateCommitDecision = {
-  disposition: "hard-navigate";
-  trace: NavigationTrace;
-};
-type NoCommitDecision = {
-  disposition: "no-commit";
-  trace: NavigationTrace;
-};
-type CommitDecision = VisibleCommitDecision | HardNavigateCommitDecision | NoCommitDecision;
-const approvedVisibleCommitBrand: unique symbol = Symbol("ApprovedVisibleCommit");
-export type ApprovedVisibleCommit = {
-  readonly [approvedVisibleCommitBrand]: true;
-  readonly action: AppRouterAction;
-  readonly decision: VisibleCommitDecision;
-  readonly interceptionContext: string | null;
-  readonly previousNextUrl: string | null;
-  readonly rootLayoutTreePath: string | null;
-  readonly routeId: string;
-};
-type VisibleCommitApproval = {
-  approvedCommit: ApprovedVisibleCommit;
-  decision: VisibleCommitDecision;
-};
-type NonVisibleCommitApproval = {
-  approvedCommit: null;
-  decision: HardNavigateCommitDecision | NoCommitDecision;
-};
-type CommitApproval = VisibleCommitApproval | NonVisibleCommitApproval;
-type ClassifiedPendingNavigationCommit = {
-  approvedCommit: ApprovedVisibleCommit | null;
-  decision: CommitDecision;
-  disposition: PendingNavigationCommitDisposition;
-  pending: PendingNavigationCommit;
   trace: NavigationTrace;
 };
 
@@ -300,13 +261,6 @@ export function routerReducer(state: AppRouterState, action: AppRouterAction): A
   }
 }
 
-export function applyApprovedVisibleCommit(
-  state: AppRouterState,
-  commit: ApprovedVisibleCommit,
-): AppRouterState {
-  return routerReducer(state, commit.action);
-}
-
 export function shouldHardNavigate(
   currentRootLayoutTreePath: string | null,
   nextRootLayoutTreePath: string | null,
@@ -338,34 +292,6 @@ export function resolvePendingNavigationCommitDisposition(options: {
   return "dispatch";
 }
 
-function resolvePendingNavigationCommitDecision(options: {
-  activeNavigationId: number;
-  currentRootLayoutTreePath: string | null;
-  nextRootLayoutTreePath: string | null;
-  startedNavigationId: number;
-}): CommitDecision {
-  const { disposition, trace } = resolvePendingNavigationCommitDispositionDecision(options);
-
-  switch (disposition) {
-    case "skip":
-      return { disposition: "no-commit", trace };
-    case "hard-navigate":
-      return { disposition: "hard-navigate", trace };
-    case "dispatch":
-      return createVisibleCommitDecision(trace);
-    default: {
-      const _exhaustive: never = disposition;
-      throw new Error("[vinext] Unknown navigation commit disposition: " + String(_exhaustive));
-    }
-  }
-}
-
-function createVisibleCommitDecision(
-  trace: NavigationTrace = createNavigationTrace(NavigationTraceReasonCodes.commitCurrent),
-): VisibleCommitDecision {
-  return { disposition: "commit", trace };
-}
-
 export function resolvePendingNavigationCommitDispositionDecision(options: {
   activeNavigationId: number;
   currentRootLayoutTreePath: string | null;
@@ -391,67 +317,6 @@ export function resolvePendingNavigationCommitDispositionDecision(options: {
       traceFields,
     ),
   };
-}
-
-function createApprovedVisibleCommit(options: {
-  decision: VisibleCommitDecision;
-  pending: PendingNavigationCommit;
-}): ApprovedVisibleCommit {
-  return {
-    [approvedVisibleCommitBrand]: true,
-    action: options.pending.action,
-    decision: options.decision,
-    interceptionContext: options.pending.interceptionContext,
-    previousNextUrl: options.pending.previousNextUrl,
-    rootLayoutTreePath: options.pending.rootLayoutTreePath,
-    routeId: options.pending.routeId,
-  };
-}
-
-export function approveHmrVisibleCommit(pending: PendingNavigationCommit): ApprovedVisibleCommit {
-  if (pending.action.operation.lane !== "hmr") {
-    throw new Error("[vinext] HMR visible commit approval requires an HMR pending operation");
-  }
-
-  return createApprovedVisibleCommit({
-    decision: createVisibleCommitDecision(),
-    pending,
-  });
-}
-
-export function approvePendingNavigationCommit(options: {
-  activeNavigationId: number;
-  currentState: AppRouterState;
-  pending: PendingNavigationCommit;
-  startedNavigationId: number;
-}): CommitApproval {
-  const decision = resolvePendingNavigationCommitDecision({
-    activeNavigationId: options.activeNavigationId,
-    currentRootLayoutTreePath: options.currentState.rootLayoutTreePath,
-    nextRootLayoutTreePath: options.pending.rootLayoutTreePath,
-    startedNavigationId: options.startedNavigationId,
-  });
-
-  switch (decision.disposition) {
-    case "commit":
-      return {
-        approvedCommit: createApprovedVisibleCommit({
-          decision,
-          pending: options.pending,
-        }),
-        decision,
-      };
-    case "hard-navigate":
-    case "no-commit":
-      return {
-        approvedCommit: null,
-        decision,
-      };
-    default: {
-      const _exhaustive: never = decision;
-      throw new Error("[vinext] Unknown commit decision: " + String(_exhaustive));
-    }
-  }
 }
 
 function getPendingNavigationCommitDispositionTraceCode(options: {
@@ -514,58 +379,4 @@ export async function createPendingNavigationCommit(options: {
     rootLayoutTreePath: metadata.rootLayoutTreePath,
     routeId: metadata.routeId,
   };
-}
-
-export async function resolveAndClassifyNavigationCommit(options: {
-  activeNavigationId: number;
-  currentState: AppRouterState;
-  navigationSnapshot: ClientNavigationRenderSnapshot;
-  nextElements: Promise<AppElements>;
-  operationLane: OperationLane;
-  previousNextUrl?: string | null;
-  renderId: number;
-  startedNavigationId: number;
-  type: "navigate" | "replace" | "traverse";
-}): Promise<ClassifiedPendingNavigationCommit> {
-  const pending = await createPendingNavigationCommit({
-    currentState: options.currentState,
-    nextElements: options.nextElements,
-    navigationSnapshot: options.navigationSnapshot,
-    operationLane: options.operationLane,
-    previousNextUrl: options.previousNextUrl,
-    renderId: options.renderId,
-    type: options.type,
-  });
-
-  const approval = approvePendingNavigationCommit({
-    activeNavigationId: options.activeNavigationId,
-    currentState: options.currentState,
-    pending,
-    startedNavigationId: options.startedNavigationId,
-  });
-
-  return {
-    approvedCommit: approval.approvedCommit,
-    decision: approval.decision,
-    disposition: commitDecisionToPendingNavigationCommitDisposition(approval.decision),
-    pending,
-    trace: approval.decision.trace,
-  };
-}
-
-function commitDecisionToPendingNavigationCommitDisposition(
-  decision: CommitDecision,
-): PendingNavigationCommitDisposition {
-  switch (decision.disposition) {
-    case "commit":
-      return "dispatch";
-    case "hard-navigate":
-      return "hard-navigate";
-    case "no-commit":
-      return "skip";
-    default: {
-      const _exhaustive: never = decision;
-      throw new Error("[vinext] Unknown commit decision: " + String(_exhaustive));
-    }
-  }
 }
